requirements.txt
docker
PyYAML
libtmux